pear-backend/requirements.txt
fastapi
uvicorn
pydantic[email]
passlib[bcrypt]
mysql-connector-python